Move all gradle plugin code to one module
