+++ gradle-plugin/src/main/kotlin/kotlinx/rpc/compilerPlugins.kt
